edit: remove simpletokenwithok
