+++ veles/secrets/gcpapikey/detector.go
@@ -32,8 +32,8 @@ func NewDetector() veles.Detector {
 	return simpletoken.Detector{
 		MaxLen: maxTokenLength,
 		Re:     keyRe,
-		FromMatch: func(b []byte) veles.Secret {
-			return GCPAPIKey{Key: string(b)}
+		FromMatch: func(b []byte) (veles.Secret, bool) {
+			return GCPAPIKey{Key: string(b)}, true
 		},
 	}
 }
