+++ veles/secrets/gcpexpressmode/detector.go
@@ -34,8 +34,8 @@ func NewDetector() veles.Detector {
 	return &simpletoken.Detector{
 		Re:     keyRe,
 		MaxLen: maxKeyLen,
-		FromMatch: func(b []byte) veles.Secret {
-			return APIKey{Key: string(b)}
+		FromMatch: func(b []byte) (veles.Secret, bool) {
+			return APIKey{Key: string(b)}, true
 		},
 	}
 }
