+++ veles/secrets/grokxaiapikey/detector.go
@@ -56,8 +56,8 @@ func NewAPIKeyDetector() veles.Detector {
 	return simpletoken.Detector{
 		MaxLen: apiKeyMaxLen,
 		Re:     apiKeyRe,
-		FromMatch: func(b []byte) veles.Secret {
-			return GrokXAIAPIKey{Key: string(b)}
+		FromMatch: func(b []byte) (veles.Secret, bool) {
+			return GrokXAIAPIKey{Key: string(b)}, true
 		},
 	}
 }
@@ -67,8 +67,8 @@ func NewManagementKeyDetector() veles.Detector {
 	return simpletoken.Detector{
 		MaxLen: mgmtKeyMaxLen,
 		Re:     mgmtKeyRe,
-		FromMatch: func(b []byte) veles.Secret {
-			return GrokXAIManagementKey{Key: string(b)}
+		FromMatch: func(b []byte) (veles.Secret, bool) {
+			return GrokXAIManagementKey{Key: string(b)}, true
 		},
 	}
 }
