+++ veles/secrets/huggingfaceapikey/detector.go
@@ -37,8 +37,8 @@ func NewDetector() veles.Detector {
 	return simpletoken.Detector{
 		MaxLen: maxTokenLength,
 		Re:     keyRe,
-		FromMatch: func(b []byte) veles.Secret {
-			return HuggingfaceAPIKey{Key: string(b)}
+		FromMatch: func(b []byte) (veles.Secret, bool) {
+			return HuggingfaceAPIKey{Key: string(b)}, true
 		},
 	}
 }
