+++ veles/secrets/rubygemsapikey/detector.go
@@ -32,8 +32,8 @@ func NewDetector() veles.Detector {
 	return simpletoken.Detector{
 		MaxLen: maxTokenLength,
 		Re:     keyRe,
-		FromMatch: func(b []byte) veles.Secret {
-			return RubyGemsAPIKey{Key: string(b)}
+		FromMatch: func(b []byte) (veles.Secret, bool) {
+			return RubyGemsAPIKey{Key: string(b)}, true
 		},
 	}
 }
